feat: add initial Delete Button

+++ src/app/_actions/delete-post.ts
@@ -1,21 +1,16 @@
 "use server"
 
+import { getServerSession } from "next-auth"
 import { revalidatePath } from "next/cache"
-import { toast } from "sonner"
 
-interface DeletePageProps {
-  id: string
-}
-
-export const deletePost = async (id: DeletePageProps) => {
-  try {
-    await fetch(`http://localhost:3108/posts/${id}`, {
-      method: "DELETE",
-    })
-    toast.success("Post apagado com sucesso")
-  } catch (error) {
-    console.error(error)
-    toast.error("Erro ao apagar post")
+export const deletePost = async (id: string) => {
+  const session = await getServerSession()
+  if (!session) {
+    return
   }
-  revalidatePath("/posts")
+
+  await fetch(`http://localhost:3108/posts/${id}`, {
+    method: "DELETE",
+  })
+  revalidatePath("/private/posts-admin")
 }
